initial seperation
wip

wip

wip

wip

wip

wip
