--- authentik/root/storages.py
@@ -1,1765 +1,0 @@
-"""Storage backends for authentik with multi-tenant support.
-
-This module provides custom storage backends for handling file storage in a multi-tenant
-environment. It supports both filesystem and S3 storage options with proper tenant isolation.
-"""
-
-import io
-import os
-import uuid
-from pathlib import Path
-from urllib.parse import parse_qs, urlencode, urlparse, urlunparse
-
-import boto3
-from botocore.config import Config
-from botocore.exceptions import ClientError, NoCredentialsError, NoRegionError
-from defusedxml import ElementTree
-from django.conf import settings
-from django.core.exceptions import ImproperlyConfigured, SuspiciousOperation
-from django.core.files.storage import FileSystemStorage
-from django.core.files.uploadedfile import UploadedFile
-from django.db import connection
-from PIL import Image
-from storages.backends.s3 import S3Storage as BaseS3Storage
-from storages.utils import safe_join
-from structlog.stdlib import get_logger
-
-from authentik.lib.config import CONFIG
-
-LOGGER = get_logger()
-
-# Storage subdirectories
-STORAGE_DIR_APPLICATION_ICONS = "application-icons"
-STORAGE_DIR_SOURCE_ICONS = "source-icons"
-STORAGE_DIR_FLOW_BACKGROUNDS = "flow-backgrounds"
-STORAGE_DIR_PUBLIC = "public"
-
-STORAGE_DIRS = [
-    STORAGE_DIR_APPLICATION_ICONS,
-    STORAGE_DIR_SOURCE_ICONS,
-    STORAGE_DIR_FLOW_BACKGROUNDS,
-    STORAGE_DIR_PUBLIC,
-]
-
-# Mapping of allowed file extensions to their corresponding MIME types
-ALLOWED_IMAGE_EXTENSIONS = {
-    ".jpg": "image/jpeg",
-    ".jpeg": "image/jpeg",
-    ".png": "image/png",
-    ".gif": "image/gif",
-    ".webp": "image/webp",
-    ".svg": "image/svg+xml",
-    ".ico": "image/x-icon",
-}
-
-
-def _validate_svg_content(content: str) -> bool:
-    """Validate SVG content structure.
-
-    Args:
-        content: SVG content as string
-
-    Returns:
-        bool: True if content is valid SVG, False otherwise
-    """
-    try:
-        # Basic check for SVG element presence
-        has_svg_element = "<svg" in content and "</svg>" in content
-        if not has_svg_element:
-            LOGGER.warning("Missing SVG element or closing tag")
-            return False
-
-        # Try to parse as XML to validate structure
-        tree = ElementTree.fromstring(content.encode())
-
-        # Validate that the root element or a child is an SVG element
-        if tree.tag.lower().endswith("svg"):
-            return True
-
-        for child in tree:
-            if child.tag.lower().endswith("svg"):
-                return True
-
-        LOGGER.warning("SVG element not found in XML structure")
-        return False
-    except ElementTree.ParseError as e:
-        LOGGER.warning("Invalid SVG XML structure", error=str(e))
-        return False
-    except ValueError as e:
-        LOGGER.warning("Invalid SVG content", error=str(e))
-        return False
-    except Exception as e:
-        LOGGER.warning("Unexpected error validating SVG", error=str(e))
-        return False
-
-
-def _validate_ico_content(content: bytes) -> bool:
-    """Validate ICO file content.
-
-    Args:
-        content: ICO file content as bytes
-
-    Returns:
-        bool: True if content is valid ICO, False otherwise
-    """
-    # ICO files should start with the magic number 0x00 0x00 0x01 0x00
-    # but we don't need to check the exact content - just the header
-    ICO_HEADER_SIZE = 4
-    return len(content) >= ICO_HEADER_SIZE and content.startswith(b"\x00\x00\x01\x00")
-
-
-def _validate_pillow_image(file: UploadedFile, ext: str, name: str = "") -> bool:
-    """Validate image using Pillow.
-
-    Args:
-        file: Uploaded file
-        ext: File extension
-        name: Name of the file for logging purposes
-
-    Returns:
-        bool: True if file is valid image, False otherwise
-    """
-    try:
-        with Image.open(file) as img:
-            format_to_ext = {
-                "JPEG": ".jpg",
-                "PNG": ".png",
-                "GIF": ".gif",
-                "WEBP": ".webp",
-            }
-            detected_ext = format_to_ext.get(img.format)
-
-            if not detected_ext:
-                LOGGER.warning("Unrecognized image format", format=img.format, extension=ext)
-                return False
-
-            # Special handling for JPEG extension variants
-            is_jpeg = detected_ext == ".jpg" and ext in (".jpg", ".jpeg")
-            if not (detected_ext == ext or is_jpeg):
-                LOGGER.warning(
-                    "File extension doesn't match content",
-                    detected_format=img.format,
-                    extension=ext,
-                )
-                return False
-
-            # Verify image data integrity
-            img.verify()
-            return True
-
-    except Exception as e:
-        LOGGER.warning("Image validation failed", error=str(e), name=name)
-        raise FileValidationError(f"Failed to validate image: {str(e)}", status_code=415) from e
-    finally:
-        file.seek(0)
-
-
-class FileValidationError(SuspiciousOperation):
-    """Custom exception for file validation errors with status code and user message."""
-
-    def __init__(self, message: str, status_code: int = 400):
-        super().__init__(message)
-        self.status_code = status_code
-        self.user_message = message
-
-
-def validate_image_file(file: UploadedFile) -> bool:
-    """Validate that the uploaded file is a valid image in an allowed format.
-
-    Args:
-        file: The uploaded file to validate
-
-    Returns:
-        bool: True if file is valid
-
-    Raises:
-        FileValidationError: If file validation fails with specific error message and status code
-    """
-    if not file:
-        raise FileValidationError("No file was provided", status_code=400)
-
-    if not hasattr(file, "content_type") or not hasattr(file, "name"):
-        raise FileValidationError("File type could not be determined", status_code=400)
-
-    name = file.name.lower() if file.name else ""
-    ext = os.path.splitext(name)[1] if name else ""
-
-    # Check if extension is allowed
-    if ext not in ALLOWED_IMAGE_EXTENSIONS:
-        allowed_exts = ", ".join(ALLOWED_IMAGE_EXTENSIONS.keys())
-        raise FileValidationError(
-            f"File type '{ext}' is not allowed. Allowed types are: {allowed_exts}",
-            status_code=415,  # Unsupported Media Type
-        )
-
-    # Check content type
-    expected_type = ALLOWED_IMAGE_EXTENSIONS.get(ext)
-    if file.content_type != expected_type:
-        raise FileValidationError(
-            f"Invalid content type '{file.content_type}' for {ext} file. Expected: {expected_type}",
-            status_code=415,
-        )
-
-    # Validate file content based on type
-    try:
-        if ext == ".svg":
-            content = file.read().decode("utf-8")
-            file.seek(0)  # Reset file position
-            if not _validate_svg_content(content):
-                raise FileValidationError("Invalid SVG file format", status_code=415)
-        elif ext == ".ico":
-            content = file.read()
-            file.seek(0)  # Reset file position
-            if not _validate_ico_content(content):
-                raise FileValidationError("Invalid ICO file format", status_code=415)
-        else:
-            # For other image types, use Pillow validation
-            try:
-                with Image.open(file) as img:
-                    # Verify image data integrity
-                    img.verify()
-                    # Reset file position after verify
-                    file.seek(0)
-            except Exception as e:
-                raise FileValidationError(f"Invalid image format: {str(e)}", status_code=415) from e
-
-        return True
-    except FileValidationError:
-        # Re-raise FileValidationError exceptions
-        raise
-    except Exception as e:
-        LOGGER.warning("Unexpected error in image validation", error=str(e), name=name)
-        raise FileValidationError(f"Failed to validate image: {str(e)}", status_code=415) from e
-
-
-def optimize_image(content):
-    """Optimize image by resizing if needed and applying compression.
-
-    Used for application icons and other image uploads to reduce file size
-    and improve loading performance.
-
-    Args:
-        content: File content to optimize (must be an image)
-
-    Returns:
-        Optimized content or original content if optimization failed
-    """
-    if not hasattr(content, "content_type") or not content.content_type.startswith("image/"):
-        return content
-
-    # Skip for SVG and ICO files which don't support Pillow optimization
-    name = content.name.lower() if hasattr(content, "name") else ""
-    ext = os.path.splitext(name)[1] if name else ""
-    if ext in (".svg", ".ico"):
-        return content
-
-    original_pos = content.tell() if hasattr(content, "tell") else 0
-
-    try:
-        # Try to open the image
-        img = Image.open(content)
-
-        # Reset file position after reading
-        if hasattr(content, "seek"):
-            content.seek(0)
-
-        # Check if we need to optimize this image
-        if img.format not in ("JPEG", "PNG", "GIF", "WEBP"):
-            return content
-
-        # Create in-memory buffer for the optimized image
-        buffer = io.BytesIO()
-
-        # Maximum dimension for images in pixels
-        MAX_IMAGE_DIMENSION = 512
-
-        # Resize large images to a reasonable size
-        if max(img.size) > MAX_IMAGE_DIMENSION:
-            LOGGER.debug(
-                "Resizing large image", original_size=img.size, format=img.format, name=name
-            )
-            ratio = float(MAX_IMAGE_DIMENSION) / max(img.size)
-            new_size = (int(img.size[0] * ratio), int(img.size[1] * ratio))
-            img = img.resize(new_size, Image.LANCZOS)
-
-        # Save with optimization based on format
-        if img.format == "JPEG":
-            img.save(buffer, format="JPEG", quality=85, optimize=True)
-        elif img.format == "PNG":
-            img.save(buffer, format="PNG", optimize=True)
-        elif img.format == "GIF":
-            img.save(buffer, format="GIF", optimize=True)
-        elif img.format == "WEBP":
-            img.save(buffer, format="WEBP", quality=85, method=4)
-        else:
-            # Fallback for unsupported optimization
-            return content
-
-        # Reset buffer position
-        buffer.seek(0)
-
-        # Create a new ContentFile with the optimized image
-        from django.core.files.base import ContentFile
-
-        optimized = ContentFile(buffer.getvalue())
-
-        # Copy needed attributes from original content
-        optimized.name = content.name if hasattr(content, "name") else "optimized.img"
-        optimized.content_type = content.content_type if hasattr(content, "content_type") else None
-
-        # Log the optimization results
-        if hasattr(content, "size"):
-            original_size = content.size
-            new_size = len(buffer.getvalue())
-            reduction = (1 - (new_size / original_size)) * 100 if original_size > 0 else 0
-            LOGGER.info(
-                "Image optimized",
-                original_size=original_size,
-                new_size=new_size,
-                reduction_percent=f"{reduction:.1f}%",
-                format=img.format,
-                name=name,
-            )
-
-        return optimized
-    except Exception as e:
-        LOGGER.warning("Image optimization failed, using original image", error=str(e), name=name)
-        # Reset file position on optimization failure
-        if hasattr(content, "seek"):
-            content.seek(original_pos)
-        return content
-
-
-class TenantAwareStorage:
-    """Mixin providing tenant-aware path functionality for storage backends."""
-
-    @property
-    def tenant_prefix(self) -> str:
-        """Get current tenant schema prefix.
-
-        Returns:
-            str: The current tenant's schema name from the database connection.
-        """
-        return connection.schema_name
-
-    @tenant_prefix.setter
-    def tenant_prefix(self, value):
-        """Setter for tenant_prefix property.
-
-        This is required for tests that need to set tenant-specific resources.
-        It does nothing as the property always returns connection.schema_name,
-        but it's needed to prevent AttributeError.
-        """
-        # This is a no-op, but prevents AttributeError in tests
-        # as the property is derived from the connection
-        pass
-
-    @tenant_prefix.deleter
-    def tenant_prefix(self):
-        """Deleter for tenant_prefix property.
-
-        This is required for tests that need to clean up tenant-specific resources.
-        """
-        # No-op deleter as the tenant_prefix is derived from the connection
-        pass
-
-    def get_tenant_path(self, name: str) -> str:
-        """Get tenant-specific path for storage.
-
-        Args:
-            name (str): Original file path/name.
-
-        Returns:
-            str: Path prefixed with tenant identifier for proper isolation.
-        """
-        return f"{self.tenant_prefix}/{name}"
-
-
-class DirectoryStructureMixin:
-    """Mixin providing directory structure creation functionality for storage backends."""
-    
-    def _ensure_directory_structure(self):
-        """Ensure required directory structure exists.
-        
-        This is an abstract method that should be implemented by storage backends
-        to create their required directory structure.
-        """
-        raise NotImplementedError("Storage backend must implement _ensure_directory_structure")
-
-
-class FileStorage(TenantAwareStorage, DirectoryStructureMixin, FileSystemStorage):
-    """Multi-tenant filesystem storage backend."""
-
-    def __init__(self, *args, **kwargs):
-        """Initialize the storage backend with tenant-aware configuration.
-
-        Creates the base storage directory if it doesn't exist and sets up proper
-        permissions and logging.
-
-        Args:
-            *args: Variable length argument list passed to parent classes
-            **kwargs: Arbitrary keyword arguments passed to parent classes
-
-        Raises:
-            PermissionError: If storage directory cannot be created due to permissions
-            OSError: If storage directory cannot be created due to filesystem errors
-        """
-        super().__init__(*args, **kwargs)
-        # Initialize _base_path from kwargs or settings
-        self._base_path = Path(kwargs.get("location", settings.MEDIA_ROOT))
-        try:
-            # Ensure the base directory exists with correct permissions
-            os.makedirs(self._base_path, exist_ok=True)
-
-            # Also create tenant-specific directory
-            tenant_dir = self._base_path / self.tenant_prefix
-            os.makedirs(tenant_dir, exist_ok=True)
-
-            # Create standard subdirectories
-            self._ensure_directory_structure()
-
-            LOGGER.debug(
-                "Storage directories initialized",
-                base_path=str(self._base_path),
-                tenant_dir=str(tenant_dir),
-                pid=os.getpid(),
-                schema_name=connection.schema_name,
-                domain_url=None,
-            )
-        except PermissionError as e:
-            LOGGER.error(
-                "Permission error creating storage directory",
-                path=str(self._base_path),
-                error=str(e),
-            )
-            raise PermissionError(
-                f"Cannot create storage directory '{self._base_path}'. Permission denied."
-            ) from e
-        except OSError as e:
-            LOGGER.error(
-                "OS error creating storage directory", path=str(self._base_path), error=str(e)
-            )
-            raise OSError(
-                f"Cannot create storage directory '{self._base_path}'. System error: {str(e)}"
-            ) from e
-
-    def _ensure_directory_structure(self):
-        """Ensure required directory structure exists in the filesystem.
-        
-        Creates all required directories for the current tenant with proper permissions.
-        """
-        try:
-            LOGGER.debug("Ensuring filesystem directory structure exists", tenant=self.tenant_prefix)
-            
-            # Create each required directory
-            for directory in STORAGE_DIRS:
-                dir_path = self._base_path / self.tenant_prefix / directory
-                try:
-                    os.makedirs(dir_path, exist_ok=True)
-                    # Create a .keep file to ensure the directory is tracked by git
-                    keep_file = dir_path / ".keep"
-                    if not keep_file.exists():
-                        keep_file.touch()
-                    LOGGER.debug("Created directory", directory=directory, path=str(dir_path), tenant=self.tenant_prefix)
-                except OSError as e:
-                    LOGGER.error(
-                        "Failed to create directory",
-                        directory=directory,
-                        path=str(dir_path),
-                        error=str(e),
-                        tenant=self.tenant_prefix,
-                    )
-                    raise PermissionError(f"Cannot create directory '{dir_path}': {str(e)}") from e
-                    
-            LOGGER.debug("Filesystem directory structure verified", tenant=self.tenant_prefix)
-            
-        except Exception as e:
-            LOGGER.error("Unexpected error creating directory structure", error=str(e), tenant=self.tenant_prefix)
-            raise OSError(f"Failed to create directory structure: {str(e)}") from e
-
-    def get_valid_name(self, name: str) -> str:
-        """Return a sanitized filename safe for storage.
-
-        Removes path components and applies additional sanitization from parent class.
-
-        Args:
-            name (str): Original filename
-
-        Returns:
-            str: Sanitized filename safe for storage
-        """
-        name = os.path.basename(name)
-        return super().get_valid_name(name)
-
-    @property
-    def base_location(self) -> Path:
-        """Get base storage directory including tenant prefix.
-
-        Returns:
-            Path: Complete path to tenant-specific storage directory
-        """
-        return Path(self._base_path) / self.tenant_prefix
-
-    @property
-    def location(self) -> str:
-        """Get absolute path to storage directory.
-
-        Returns:
-            str: Absolute filesystem path to tenant storage directory
-        """
-        return os.path.abspath(self.base_location)
-
-    @property
-    def base_url(self) -> str:
-        """Get base URL for serving stored files with tenant prefix.
-
-        Ensures proper URL composition by validating and fixing MEDIA_URL format.
-
-        Returns:
-            str: Base URL with proper tenant prefix for serving files
-        """
-        base_url = settings.MEDIA_URL
-        if not base_url.endswith("/"):
-            LOGGER.warning(
-                "MEDIA_URL should end with '/' for proper URL composition", current_value=base_url
-            )
-            base_url += "/"
-        return f"{base_url}{self.tenant_prefix}/"
-
-    def _validate_path(self, name: str) -> str:
-        """Validate the path for security issues.
-
-        Ensures that the path does not contain suspicious characters or attempt to
-        traverse outside the storage directory.
-
-        Args:
-            name (str): Name of the file to validate
-
-        Returns:
-            str: Validated path name
-
-        Raises:
-            SuspiciousOperation: If the path contains invalid characters or traversal attempts
-        """
-        if not name:
-            raise SuspiciousOperation("Empty filename is not allowed")
-
-        # Check for directory traversal attempts
-        if ".." in name.split("/") or ".." in name.split("\\"):
-            raise SuspiciousOperation(f"Invalid characters in filename '{name}'")
-
-        # Convert to posix path and normalize
-        clean_name = str(Path(name).as_posix())
-
-        # Ensure the path is relative and doesn't start with / or other special patterns
-        while clean_name.startswith("/"):
-            clean_name = clean_name[1:]
-
-        # Final validation using safe_join
-        try:
-            # We use safe_join for final validation
-            safe_join("", clean_name)
-        except ValueError as e:
-            raise SuspiciousOperation(f"Invalid characters in filename '{name}'") from e
-
-        return clean_name
-
-    def path(self, name: str) -> str:
-        """Return the absolute path to the file.
-
-        Args:
-            name (str): The name of the file including tenant prefix
-
-        Returns:
-            str: The absolute path to the file on the filesystem
-
-        Raises:
-            SuspiciousOperation: If the file path attempts to traverse outside the storage directory
-        """
-        # Apply tenant prefix if not already included in the name
-        if not name.startswith(f"{self.tenant_prefix}/"):
-            tenant_path = self.get_tenant_path(name)
-        else:
-            tenant_path = name
-
-        # Normalize the path to prevent path traversal
-        clean_name = self._validate_path(tenant_path)
-
-        # Join the base location with the validated name
-        return os.path.join(self.location, clean_name.replace(f"{self.tenant_prefix}/", "", 1))
-
-    def _get_file_subdirectory(self, name: str, content_type: str | None = None) -> str:
-        """Determine the appropriate subdirectory for a file based on its name and content type.
-
-        Args:
-            name (str): Original filename
-            content_type (str | None): Content type of the file if available
-
-        Returns:
-            str: Subdirectory path where the file should be stored
-        """
-        name_lower = name.lower()
-
-        # Application icons
-        if any(x in name_lower for x in ["app-icon", "application-icon"]):
-            return STORAGE_DIR_APPLICATION_ICONS
-
-        # Source icons
-        if any(x in name_lower for x in ["source-icon", "source-logo"]):
-            return STORAGE_DIR_SOURCE_ICONS
-
-        # Flow backgrounds
-        if any(x in name_lower for x in ["flow-bg", "flow-background"]):
-            return STORAGE_DIR_FLOW_BACKGROUNDS
-
-        # Default to public for other files
-        return STORAGE_DIR_PUBLIC
-
-    def _save(self, name: str, content) -> str:
-        """Save the file with content validation and tenant prefix application.
-
-        Args:
-            name (str): Name of the file
-            content: File content to save
-
-        Returns:
-            str: Name of the saved file with tenant prefix
-
-        Raises:
-            FileValidationError: If file validation fails (for images)
-        """
-        # First check if this is an image upload that needs validation
-        if hasattr(content, "content_type") and content.content_type.startswith("image/"):
-            try:
-                validate_image_file(content)
-                # Optimize image after validation
-                content = optimize_image(content)
-            except FileValidationError as e:
-                LOGGER.warning("Image validation failed", name=name, error=str(e))
-                raise
-
-        # For application icons, always validate
-        if name.startswith(STORAGE_DIR_APPLICATION_ICONS):
-            try:
-                validate_image_file(content)
-            except FileValidationError as e:
-                LOGGER.warning("Application icon validation failed", name=name, error=str(e))
-                raise
-
-        # Preserve the original directory structure
-        original_dir = os.path.dirname(name)
-        base_name, ext = os.path.splitext(os.path.basename(name))
-        unique_id = str(uuid.uuid4())
-        randomized_name = f"{unique_id}{ext}"
-
-        # Get appropriate subdirectory
-        subdirectory = self._get_file_subdirectory(name, getattr(content, "content_type", None))
-
-        # Create symlink directory structure
-        if original_dir:
-            symlink_dir = os.path.join(self.path(self.get_tenant_path(original_dir)))
-            os.makedirs(symlink_dir, exist_ok=True)
-
-        # Combine subdirectory with randomized name for actual storage
-        full_path = f"{subdirectory}/{randomized_name}"
-
-        # Apply tenant prefix if it's not already there
-        if not full_path.startswith(f"{self.tenant_prefix}/"):
-            tenant_path = self.get_tenant_path(full_path)
-        else:
-            tenant_path = full_path
-
-        # Perform regular file save
-        file_path = self.path(tenant_path)
-
-        # Ensure the directory exists
-        directory = os.path.dirname(file_path)
-        os.makedirs(directory, exist_ok=True)
-
-        LOGGER.debug("Saving file", name=name, path=file_path)
-
-        # Save the file in the storage location
-        saved_name = super()._save(tenant_path, content)
-
-        # Create a symlink from the original path to the stored file
-        if original_dir:
-            symlink_path = os.path.join(symlink_dir, os.path.basename(name))
-            if os.path.lexists(symlink_path):
-                os.unlink(symlink_path)
-            os.symlink(os.path.abspath(self.path(saved_name)), symlink_path)
-
-        return saved_name
-
-
-class S3StorageError(Exception):
-    """Base exception class for S3 storage errors."""
-    def __init__(self, message: str, original_error: Exception = None):
-        self.message = message
-        self.original_error = original_error
-        super().__init__(self.message)
-
-class S3BucketError(S3StorageError):
-    """Exception raised for bucket-related errors."""
-    pass
-
-class S3AccessError(S3StorageError):
-    """Exception raised for access-related errors."""
-    pass
-
-class S3UploadError(S3StorageError):
-    """Exception raised for upload-related errors."""
-    pass
-
-class S3StorageNotConfiguredError(S3StorageError):
-    """Exception raised when S3 storage is not properly configured."""
-    pass
-
-
-class S3Storage(TenantAwareStorage, DirectoryStructureMixin, BaseS3Storage):
-    """Multi-tenant S3 (compatible/Amazon) storage backend."""
-
-    CONFIG_KEYS = {
-        "session_profile": "storage.media.s3.session_profile",
-        "access_key": "storage.media.s3.access_key",
-        "secret_key": "storage.media.s3.secret_key",
-        "security_token": "storage.media.s3.security_token",
-        "bucket_name": "storage.media.s3.bucket_name",
-        "region_name": "storage.media.s3.region_name",
-        "endpoint_url": "storage.media.s3.endpoint",
-        "custom_domain": "storage.media.s3.custom_domain",
-    }
-
-    def __init__(self, **kwargs):
-        """Initialize S3Storage with configuration.
-
-        Args:
-            **kwargs: Configuration options passed to parent S3Storage
-
-        Raises:
-            S3StorageNotConfiguredError: If storage is not properly configured
-            S3BucketError: If bucket doesn't exist or cannot be accessed
-            S3AccessError: If credentials are invalid or access is denied
-        """
-        try:
-            # Initialize client/bucket references
-            self._client = None
-            self._s3_client = None
-            self._bucket = None
-
-            # Pre-fetch configuration values
-            self._session_profile = self._get_config_value("session_profile")
-            self._access_key = self._get_config_value("access_key")
-            self._secret_key = self._get_config_value("secret_key")
-            self._security_token = self._get_config_value("security_token")
-            self._bucket_name = self._get_config_value("bucket_name")
-            self._region_name = self._get_config_value("region_name")
-            self._endpoint_url = self._get_config_value("endpoint_url")
-            self._custom_domain = self._get_config_value("custom_domain")
-
-            # Debug logging
-            LOGGER.debug(
-                "S3Storage initialization",
-                has_session_profile=bool(self._session_profile),
-                has_access_key=bool(self._access_key),
-                has_secret_key=bool(self._secret_key),
-                has_security_token=bool(self._security_token),
-                bucket_name=self._bucket_name,
-                region_name=self._region_name,
-                endpoint_url=self._endpoint_url,
-                custom_domain=self._custom_domain,
-                tenant=getattr(self, "tenant_prefix", "unknown"),
-            )
-
-            # Validate configuration before proceeding
-            self._validate_configuration()
-
-            # Update kwargs with our configuration values
-            settings = kwargs.copy()
-            settings.update(
-                {
-                    "session_profile": self._session_profile,
-                    "access_key": self._access_key,
-                    "secret_key": self._secret_key,
-                    "security_token": self._security_token,
-                    "bucket_name": self._bucket_name,
-                    "region_name": self._region_name,
-                    "endpoint_url": self._endpoint_url,
-                    "custom_domain": self._custom_domain,
-                    "querystring_auth": True,
-                    "querystring_expire": 3600,
-                }
-            )
-
-            # Initialize parent class with cleaned settings
-            try:
-                super().__init__(**settings)
-                LOGGER.debug("S3Storage parent initialization successful")
-            except Exception as e:
-                LOGGER.error("S3Storage parent initialization failed", error=str(e))
-                raise S3StorageNotConfiguredError(f"Failed to initialize S3 storage: {str(e)}") from e
-
-            self._file_mapping = {}
-
-            # Apply transfer config if specified
-            transfer_config = CONFIG.refresh("storage.media.s3.transfer_config", None)
-            if transfer_config:
-                settings["transfer_config"] = Config(s3=transfer_config)
-                super().__init__(**settings)
-
-        except S3StorageError:
-            # Re-raise our custom exceptions
-            raise
-        except Exception as e:
-            # Catch any other unexpected errors
-            LOGGER.error("Unexpected error during S3 storage initialization", error=str(e))
-            raise S3StorageNotConfiguredError(f"Failed to initialize S3 storage: {str(e)}") from e
-
-    def _get_config_value(self, key: str) -> str | None:
-        """Get refreshed configuration value from environment.
-
-        Args:
-            key (str): Configuration key from CONFIG_KEYS
-
-        Returns:
-            str | None: Configuration value if set, None otherwise
-        """
-        return CONFIG.refresh(self.CONFIG_KEYS[key], None)
-
-    def _ensure_directory_structure(self):
-        """Ensure required directory structure exists in the S3 bucket.
-        
-        Creates empty marker files in each required directory to ensure the structure exists.
-        This is needed because S3 doesn't have real directories, and we need to ensure
-        the paths exist for future operations.
-        """
-        try:
-            LOGGER.debug("Ensuring S3 directory structure exists", tenant=self.tenant_prefix)
-            
-            # Create a marker file for each required directory
-            for directory in STORAGE_DIRS:
-                # Create a path with tenant prefix
-                marker_path = f"{self.tenant_prefix}/{directory}/.keep"
-                
-                try:
-                    # Check if marker already exists
-                    try:
-                        self._s3_client.head_object(Bucket=self._bucket_name, Key=marker_path)
-                        LOGGER.debug("Directory marker exists", directory=directory, tenant=self.tenant_prefix)
-                        continue
-                    except ClientError as e:
-                        if e.response["Error"]["Code"] != "404":
-                            raise
-                    
-                    # Create empty marker file
-                    self._s3_client.put_object(
-                        Bucket=self._bucket_name,
-                        Key=marker_path,
-                        Body=b"",
-                        ContentType="text/plain",
-                    )
-                    LOGGER.debug("Created directory marker", directory=directory, tenant=self.tenant_prefix)
-                    
-                except ClientError as e:
-                    error_code = e.response.get("Error", {}).get("Code", "Unknown")
-                    LOGGER.error(
-                        "Failed to create directory marker",
-                        directory=directory,
-                        error_code=error_code,
-                        error=str(e),
-                        tenant=self.tenant_prefix,
-                    )
-                    if error_code in ("AccessDenied", "AllAccessDisabled"):
-                        raise S3AccessError(f"No permission to create directory structure in bucket '{self._bucket_name}'") from e
-                    else:
-                        raise S3BucketError(f"Failed to create directory structure: {error_code}") from e
-                    
-            LOGGER.debug("S3 directory structure verified", tenant=self.tenant_prefix)
-            
-        except S3StorageError:
-            raise
-        except Exception as e:
-            LOGGER.error("Unexpected error creating directory structure", error=str(e), tenant=self.tenant_prefix)
-            raise S3BucketError(f"Failed to create directory structure: {str(e)}") from e
-
-    def _validate_configuration(self):
-        """Validate S3 configuration and credentials.
-
-        Checks that all required configuration keys are set and that the
-        bucket exists and is accessible.
-
-        Raises:
-            S3BucketError: If bucket doesn't exist or cannot be accessed
-            S3AccessError: If credentials are invalid or access is denied
-            S3StorageNotConfiguredError: If storage is not properly configured
-        """
-        try:
-            # Check that bucket_name and region_name are set
-            if not self._bucket_name:
-                LOGGER.error("Missing required S3 configuration: bucket_name")
-                raise S3StorageNotConfiguredError("Missing required S3 configuration: bucket_name")
-
-            if not self._region_name:
-                LOGGER.info("No region_name specified, defaulting to us-east-1")
-                self._region_name = "us-east-1"
-
-            has_profile = bool(self._session_profile)
-            has_credentials = bool(self._access_key) and bool(self._secret_key)
-
-            # Check that session profile is not provided with access key and secret key
-            if has_profile and has_credentials:
-                raise S3StorageNotConfiguredError(
-                    "AWS session profile should not be provided with access key and secret key"
-                )
-
-            if not (has_profile or has_credentials):
-                LOGGER.error(
-                    "Missing required S3 authentication configuration. "
-                    "Either session_profile or (access_key and secret_key) must be set."
-                )
-                raise S3StorageNotConfiguredError(
-                    "Missing required S3 authentication configuration. "
-                    "Either session_profile or (access_key and secret_key) must be set."
-                )
-
-            # Validate bucket exists and is accessible by attempting to list objects
-            try:
-                _ = self.client  # Ensure client is initialized
-                # Try to list objects in the bucket with max_keys=1 to minimize data transfer
-                try:
-                    self._s3_client.list_objects_v2(Bucket=self._bucket_name, MaxKeys=1)
-                except ClientError as e:
-                    error_code = e.response.get("Error", {}).get("Code", "Unknown")
-                    if error_code == "NoSuchBucket":
-                        LOGGER.error("S3 bucket does not exist", bucket=self._bucket_name)
-                        raise S3BucketError(f"S3 bucket '{self._bucket_name}' does not exist") from e
-                    elif error_code == "NoSuchKey":
-                        # NoSuchKey is not a critical error - it just means the bucket is empty
-                        LOGGER.debug("Bucket is empty", bucket=self._bucket_name)
-                    elif error_code in ("AccessDenied", "AllAccessDisabled"):
-                        LOGGER.error("No permission to access S3 bucket", bucket=self._bucket_name)
-                        raise S3AccessError(f"No permission to access S3 bucket '{self._bucket_name}'") from e
-                    elif error_code == "InvalidAccessKeyId":
-                        LOGGER.error("Invalid AWS credentials", bucket=self._bucket_name)
-                        raise S3AccessError("Invalid AWS credentials") from e
-                    elif error_code == "BucketRegionError":
-                        LOGGER.error("Bucket region mismatch", bucket=self._bucket_name)
-                        raise S3BucketError("Bucket region mismatch") from e
-                    else:
-                        LOGGER.error(
-                            "Error accessing S3 bucket",
-                            bucket=self._bucket_name,
-                            error=str(e),
-                            code=error_code,
-                        )
-                        raise S3BucketError(f"Error accessing S3 bucket: {error_code}") from e
-
-                # Ensure required directory structure exists
-                self._ensure_directory_structure()
-
-            except (NoCredentialsError, NoRegionError) as e:
-                LOGGER.error("AWS credentials/region configuration error", error=str(e))
-                raise S3StorageNotConfiguredError(f"AWS configuration error: {str(e)}") from e
-            except Exception as e:
-                LOGGER.error("Unexpected error during S3 configuration validation", error=str(e))
-                raise S3StorageNotConfiguredError(f"Unexpected error during S3 configuration: {str(e)}") from e
-
-            LOGGER.debug("S3 configuration validated successfully", bucket=self._bucket_name)
-
-        except S3StorageError:
-            # Re-raise our custom exceptions
-            raise
-        except Exception as e:
-            # Catch any other unexpected errors
-            LOGGER.error("Unexpected error during S3 configuration validation", error=str(e))
-            raise S3StorageNotConfiguredError(f"Unexpected error during S3 configuration: {str(e)}") from e
-
-    @property
-    def client(self):
-        """Get or create boto3 S3 client with current credentials.
-
-        Creates a new boto3 S3 client if none exists, using current AWS credentials.
-
-        Returns:
-            boto3.client: Configured S3 client instance
-
-        Raises:
-            ImproperlyConfigured: If AWS credentials are invalid
-            ClientError: If AWS client initialization fails
-        """
-        if not self._client or not self._s3_client:
-            try:
-                LOGGER.debug(
-                    "Creating boto3 session",
-                    profile_name=self._session_profile,
-                    has_access_key=(
-                        bool(self._access_key) and self._access_key[:4] + "..."
-                        if self._access_key
-                        else None
-                    ),
-                    has_secret_key=bool(self._secret_key),
-                    has_security_token=bool(self._security_token),
-                    tenant=self.tenant_prefix,
-                )
-
-                session = boto3.Session(
-                    profile_name=self._session_profile,
-                    aws_access_key_id=self._access_key,
-                    aws_secret_access_key=self._secret_key,
-                    aws_session_token=self._security_token,
-                )
-
-                LOGGER.debug(
-                    "Boto3 session created",
-                    available_profiles=session.available_profiles,
-                    profile_name=session.profile_name,
-                    region_name=session.region_name,
-                    tenant=self.tenant_prefix,
-                )
-
-                client_kwargs = {
-                    "region_name": self._region_name,
-                }
-                if self._endpoint_url:
-                    s3_config = Config(s3={"addressing_style": "path"})
-                    client_kwargs.update(
-                        {
-                            "endpoint_url": self._endpoint_url,
-                            "config": s3_config,
-                        }
-                    )
-                    LOGGER.debug(
-                        "Using custom S3 endpoint with path-style addressing",
-                        endpoint=self._endpoint_url,
-                        tenant=self.tenant_prefix,
-                    )
-
-                LOGGER.debug(
-                    "Creating S3 resource and client",
-                    client_kwargs=client_kwargs,
-                    tenant=self.tenant_prefix,
-                )
-
-                self._client = session.resource("s3", **client_kwargs)
-                self._s3_client = session.client("s3", **client_kwargs)
-
-                LOGGER.debug(
-                    "Created S3 resource and client",
-                    session_profile=self._session_profile,
-                    region=self._region_name,
-                    endpoint=self._endpoint_url,
-                    tenant=self.tenant_prefix,
-                )
-            except (NoCredentialsError, NoRegionError) as e:
-                LOGGER.critical(
-                    "AWS credentials/region configuration error",
-                    error=str(e),
-                    error_type=type(e).__name__,
-                    tenant=self.tenant_prefix,
-                )
-                raise ImproperlyConfigured(f"AWS configuration error: {e}") from e
-
-        return self._client
-
-    @property
-    def bucket(self):
-        """Get or create S3 bucket instance with access validation.
-
-        Creates a new S3 bucket instance if none exists and validates access permissions.
-
-        Returns:
-            boto3.s3.Bucket: Validated S3 bucket instance
-
-        Raises:
-            ImproperlyConfigured: If bucket doesn't exist or permissions are insufficient
-            ClientError: If bucket access fails
-        """
-        if not self._bucket:
-            bucket_name = self._get_config_value("bucket_name")
-            try:
-                # First check credentials by listing buckets
-                try:
-                    LOGGER.debug(
-                        "Listing S3 buckets to validate credentials",
-                        tenant=self.tenant_prefix,
-                    )
-                    buckets = list(self.client.buckets.all())
-                    bucket_names = [b.name for b in buckets]
-                    LOGGER.debug(
-                        "Successfully listed S3 buckets",
-                        bucket_count=len(bucket_names),
-                        buckets=bucket_names,
-                        target_bucket=bucket_name,
-                        bucket_exists=bucket_name in bucket_names,
-                        tenant=self.tenant_prefix,
-                    )
-                except (ClientError, NoCredentialsError) as e:
-                    if isinstance(e, ClientError):
-                        error_code = e.response.get("Error", {}).get("Code", "Unknown")
-                        error_message = e.response.get("Error", {}).get("Message", "Unknown error")
-                        LOGGER.critical(
-                            "Invalid AWS credentials",
-                            error_code=error_code,
-                            message=error_message,
-                            response=str(e.response),
-                            tenant=self.tenant_prefix,
-                        )
-                    else:
-                        LOGGER.critical(
-                            "Invalid AWS credentials",
-                            error=str(e),
-                            error_type=type(e).__name__,
-                            tenant=self.tenant_prefix,
-                        )
-                    raise ImproperlyConfigured("Invalid AWS credentials") from e
-
-                # Then check bucket existence and permissions
-                try:
-                    LOGGER.debug(
-                        "Checking S3 bucket existence and permissions",
-                        bucket=bucket_name,
-                        tenant=self.tenant_prefix,
-                    )
-                    bucket = self.client.Bucket(bucket_name)
-                    # Try to access the bucket to verify permissions
-                    list(bucket.objects.limit(1))
-                    LOGGER.debug(
-                        "Successfully verified S3 bucket access",
-                        bucket=bucket_name,
-                        tenant=self.tenant_prefix,
-                    )
-                except ClientError as e:
-                    error_code = e.response.get("Error", {}).get("Code", "Unknown")
-                    error_message = e.response.get("Error", {}).get("Message", "Unknown error")
-                    if error_code == "NoSuchBucket":
-                        LOGGER.error(
-                            "S3 bucket does not exist",
-                            bucket=bucket_name,
-                            error_code=error_code,
-                            message=error_message,
-                            tenant=self.tenant_prefix,
-                        )
-                        raise ImproperlyConfigured(
-                            f"S3 bucket '{bucket_name}' does not exist"
-                        ) from e
-                    elif error_code in ("AccessDenied", "AllAccessDisabled"):
-                        LOGGER.error(
-                            "No permission to access S3 bucket",
-                            bucket=bucket_name,
-                            error_code=error_code,
-                            message=error_message,
-                            tenant=self.tenant_prefix,
-                        )
-                        raise ImproperlyConfigured(
-                            f"No permission to access S3 bucket '{bucket_name}'"
-                        ) from e
-                    else:
-                        LOGGER.error(
-                            "Error accessing S3 bucket",
-                            bucket=bucket_name,
-                            error_code=error_code,
-                            message=error_message,
-                            response=str(e.response),
-                            tenant=self.tenant_prefix,
-                        )
-                        raise
-
-                self._bucket = bucket
-                LOGGER.info(
-                    "Successfully connected to S3 bucket",
-                    bucket=bucket_name,
-                    region=self._region_name,
-                    endpoint=self._endpoint_url,
-                    tenant=self.tenant_prefix,
-                )
-
-            except Exception as e:
-                LOGGER.error(
-                    "Unexpected error accessing S3",
-                    error=str(e),
-                    error_type=type(e).__name__,
-                    tenant=self.tenant_prefix,
-                )
-                if isinstance(e, ImproperlyConfigured):
-                    raise
-                raise ImproperlyConfigured(f"S3 configuration error: {str(e)}") from e
-
-        return self._bucket
-
-    @property
-    def base_url(self) -> str:
-        """Get base URL for S3 storage with tenant prefix.
-
-        Returns:
-            str: Base URL with tenant prefix for S3 storage
-        """
-        return f"/{self.tenant_prefix}/"
-
-    def get_valid_name(self, name: str) -> str:
-        """Return a sanitized filename safe for S3 storage.
-
-        Removes path components and applies additional sanitization.
-
-        Args:
-            name (str): Original filename
-
-        Returns:
-            str: Sanitized filename safe for S3 storage
-        """
-        # For S3, we want to preserve the directory structure
-        dir_name = os.path.dirname(name)
-        base_name = os.path.basename(name)
-        base_name = super().get_valid_name(base_name)
-        if dir_name:
-            return os.path.join(dir_name, base_name)
-        return base_name
-
-    def _randomize_filename(self, filename: str) -> str:
-        """Generate a randomized filename to prevent conflicts and overwriting.
-
-        Creates a unique filename by injecting a UUID while preserving the original
-        extension for proper file type handling.
-
-        Args:
-            filename (str): Original filename
-
-        Returns:
-            str: Randomized filename with UUID
-        """
-        if not filename:
-            raise SuspiciousOperation("Could not derive file name from empty string")
-
-        base_name, ext = os.path.splitext(os.path.basename(filename))
-        unique_id = str(uuid.uuid4())
-        randomized = f"{unique_id}_{base_name}{ext}"
-
-        LOGGER.debug("Randomized filename", original=filename, randomized=randomized)
-
-        return randomized
-
-    def _normalize_name(self, name: str) -> str:
-        """Normalize file name for S3 storage.
-
-        Ensures the name is properly prefixed with tenant prefix and doesn't
-        contain any suspicious characters that could lead to path traversal.
-
-        Args:
-            name (str): Original file name
-
-        Returns:
-            str: Normalized S3 key for the file
-
-        Raises:
-            SuspiciousOperation: If the name contains invalid characters
-        """
-        # Clean the name by removing leading slashes and normalizing to forward slashes
-        clean_name = str(Path(name).as_posix())
-        while clean_name.startswith("/"):
-            clean_name = clean_name[1:]
-
-        # Check for directory traversal attempts
-        if ".." in clean_name.split("/"):
-            raise SuspiciousOperation(f"Invalid characters in filename '{name}'")
-
-        # Final validation
-        try:
-            safe_join("", clean_name)
-        except ValueError as e:
-            raise SuspiciousOperation(f"Invalid characters in filename '{name}'") from e
-
-        # If the name doesn't already have the tenant prefix, add it
-        # Skip this for unit tests where we expect specific key names
-        if not clean_name.startswith(f"{self.tenant_prefix}/") and "/test_" not in clean_name:
-            clean_name = f"{self.tenant_prefix}/{clean_name}"
-
-        # Log normalization result
-        LOGGER.debug("Normalized file name", original=name, normalized=clean_name)
-
-        return clean_name
-
-    def _delete_previous_instance_file(self, content) -> None:
-        """Delete the previous file from the model instance if it exists."""
-        if not (hasattr(content, "_instance") and hasattr(content._instance, content._field.name)):
-            return
-
-        old_file = getattr(content._instance, content._field.name)
-        if not old_file:
-            return
-
-        try:
-            old_name = old_file.name
-            LOGGER.debug(
-                "Deleting previous file from model instance",
-                name=old_name,
-                tenant=self.tenant_prefix,
-            )
-            old_file.delete(save=False)  # Don't save the model yet
-        except Exception as e:
-            LOGGER.warning(
-                "Failed to delete old file from model instance",
-                name=old_name,
-                error=str(e),
-                error_type=type(e).__name__,
-                tenant=self.tenant_prefix,
-            )
-
-    def _delete_previous_mapped_file(self, name: str) -> None:
-        """Delete the previous file with the same name from S3 if it exists in the mapping."""
-        if name not in self._file_mapping:
-            return
-
-        old_name = self._file_mapping[name]
-        try:
-            LOGGER.debug(
-                "Deleting previous file with same name",
-                name=name,
-                old_key=old_name,
-                tenant=self.tenant_prefix,
-            )
-            self.bucket.Object(old_name).delete()
-            self._file_mapping.pop(name)
-        except Exception as e:
-            LOGGER.warning(
-                "Failed to delete old file during replacement",
-                name=name,
-                old_key=old_name,
-                error=str(e),
-                error_type=type(e).__name__,
-                tenant=self.tenant_prefix,
-            )
-
-    def _verify_upload(self, obj, normalized_name: str) -> None:
-        """Verify that the upload was successful."""
-        LOGGER.debug(
-            "Upload to S3 completed, verifying object",
-            key=normalized_name,
-            tenant=self.tenant_prefix,
-        )
-
-        try:
-            obj_data = obj.load()
-            LOGGER.debug(
-                "Successfully verified S3 upload",
-                key=normalized_name,
-                object_data=str(obj_data),
-                tenant=self.tenant_prefix,
-            )
-        except ClientError as e:
-            error_code = e.response.get("Error", {}).get("Code", "Unknown")
-            error_message = e.response.get("Error", {}).get("Message", "Unknown error")
-            LOGGER.error(
-                "Failed to verify S3 upload",
-                key=normalized_name,
-                error_code=error_code,
-                message=error_message,
-                response=str(e.response),
-                tenant=self.tenant_prefix,
-            )
-            self._cleanup_failed_upload(obj, normalized_name)
-            raise
-
-    def _cleanup_failed_upload(self, obj, normalized_name: str) -> None:
-        """Clean up a failed upload by deleting the object."""
-        try:
-            LOGGER.debug(
-                "Cleaning up failed upload",
-                key=normalized_name,
-                tenant=self.tenant_prefix,
-            )
-            obj.delete()
-        except Exception as cleanup_error:
-            LOGGER.warning(
-                "Failed to clean up after failed upload",
-                key=normalized_name,
-                error=str(cleanup_error),
-                tenant=self.tenant_prefix,
-            )
-
-    def _log_save_attempt(
-        self, name: str, randomized_name: str, normalized_name: str, content
-    ) -> None:
-        """Log information about the file being saved to S3."""
-        LOGGER.info(
-            "Saving image to S3",
-            original_name=name,
-            randomized_name=randomized_name,
-            normalized_name=normalized_name,
-            content_type=getattr(content, "content_type", None),
-            content_length=getattr(content, "size", None),
-            tenant=self.tenant_prefix,
-        )
-
-    def _log_save_success(self, normalized_name: str, name: str) -> None:
-        """Log successful file save to S3."""
-        LOGGER.debug(
-            "Image saved successfully to S3",
-            key=normalized_name,
-            original_name=name,
-            tenant=self.tenant_prefix,
-        )
-
-    def _handle_save_error(self, e: Exception, name: str, normalized_name: str) -> None:
-        """Handle and log errors during file save operation."""
-        if isinstance(e, ClientError):
-            error_code = e.response.get("Error", {}).get("Code", "Unknown")
-            error_message = e.response.get("Error", {}).get("Message", "Unknown error")
-            LOGGER.error(
-                "Error saving file to S3",
-                name=name,
-                key=normalized_name,
-                error_code=error_code,
-                message=error_message,
-                response=str(e.response),
-                tenant=self.tenant_prefix,
-            )
-            if error_code == "NoSuchBucket":
-                raise S3BucketError(f"S3 bucket '{self._bucket_name}' does not exist") from e
-            elif error_code in ("AccessDenied", "AllAccessDisabled"):
-                raise S3AccessError(f"No permission to access S3 bucket '{self._bucket_name}'") from e
-            else:
-                raise S3UploadError(f"Failed to upload file to S3: {error_code}") from e
-        else:
-            LOGGER.error(
-                "Unexpected error saving file to S3",
-                name=name,
-                key=normalized_name,
-                error=str(e),
-                error_type=type(e).__name__,
-                tenant=self.tenant_prefix,
-            )
-            raise S3UploadError(f"Failed to upload file to S3: {str(e)}") from e
-
-    def _get_file_subdirectory(self, filename: str) -> str:
-        """Get the appropriate subdirectory for a file based on its name"""
-        filename = filename.lower()
-        if "app-icon" in filename or "application-icon" in filename:
-            return STORAGE_DIR_APPLICATION_ICONS
-        if "source-icon" in filename or "source-logo" in filename:
-            return STORAGE_DIR_SOURCE_ICONS
-        if "flow-bg" in filename or "flow-background" in filename:
-            return STORAGE_DIR_FLOW_BACKGROUNDS
-        return STORAGE_DIR_PUBLIC
-
-    def _save(self, name: str, content) -> str:
-        """Save file to S3 with validation and error handling.
-
-        Args:
-            name (str): Name of the file to save
-            content: File content to save (file-like object)
-
-        Returns:
-            str: Name of the file that was saved (with tenant prefix)
-
-        Raises:
-            FileValidationError: If image validation fails
-            ClientError: If S3 upload fails
-        """
-        # Check if the file is an image by extension
-        ext = os.path.splitext(name.lower())[1] if name else ""
-
-        # Only allow image files with valid extensions
-        if ext not in ALLOWED_IMAGE_EXTENSIONS:
-            raise SuspiciousOperation("S3Storage only accepts valid image files")
-
-        # First validate content if it's an image
-        if (
-            hasattr(content, "content_type")
-            and content.content_type
-            and content.content_type.startswith("image/")
-        ):
-            try:
-                validate_image_file(content)
-                # Optimize image after validation
-                content = optimize_image(content)
-            except FileValidationError as e:
-                LOGGER.warning("Image validation failed", name=name, error=str(e))
-                raise
-
-        # Get the appropriate subdirectory
-        subdirectory = self._get_file_subdirectory(name)
-
-        # Generate a random filename with the original extension
-        random_name = f"{uuid.uuid4()}{ext}"
-
-        # Combine subdirectory and random name
-        final_name = f"{subdirectory}/{random_name}"
-
-        # Add tenant prefix
-        name = self.get_tenant_path(final_name)
-
-        # Normalize the name for S3
-        normalized_name = self._normalize_name(name)
-
-        # Log the save attempt
-        self._log_save_attempt(name, random_name, normalized_name, content)
-
-        # Get S3 object for this file
-        obj = self.bucket.Object(normalized_name)
-
-        try:
-            # Upload the file to S3
-            LOGGER.debug(
-                "Uploading file to S3",
-                key=normalized_name,
-                tenant=self.tenant_prefix,
-            )
-            upload_kwargs = {}
-            if hasattr(content, "content_type") and content.content_type:
-                upload_kwargs["ContentType"] = content.content_type
-
-            try:
-                obj.upload_fileobj(content, ExtraArgs=upload_kwargs if upload_kwargs else None)
-            except ClientError as e:
-                # Log the error
-                error_code = e.response.get("Error", {}).get("Code", "Unknown")
-                error_message = e.response.get("Error", {}).get("Message", "Unknown error")
-                LOGGER.error(
-                    "S3 upload failed with ClientError",
-                    name=name,
-                    key=normalized_name,
-                    error_code=error_code,
-                    message=error_message,
-                    response=str(e.response),
-                    tenant=self.tenant_prefix,
-                )
-                # Clean up failed upload attempts
-                self._cleanup_failed_upload(obj, normalized_name)
-                # Re-raise the ClientError directly
-                raise
-
-            # Verify the upload was successful
-            self._verify_upload(obj, normalized_name)
-
-            # Log successful save
-            self._log_save_success(name, normalized_name)
-
-            # Return the name with tenant prefix to ensure proper path reference
-            return name
-        except Exception as e:
-            # Handle exceptions that aren't ClientError (already handled above)
-            if not isinstance(e, ClientError):
-                # Clean up failed upload attempts
-                self._cleanup_failed_upload(obj, normalized_name)
-                # Handle errors based on type
-                self._handle_save_error(e, name, normalized_name)
-            # Re-raise the exception after cleanup and logging
-            raise
-
-    def delete(self, name: str) -> None:
-        """Delete file from S3 storage.
-
-        Attempts to delete the file using either the mapped normalized name
-        or by normalizing the provided name.
-
-        Args:
-            name (str): Name of the file to delete
-
-        Raises:
-            S3AccessError: If access is denied
-            S3BucketError: If bucket doesn't exist
-        """
-        try:
-            # Get normalized name from mapping or normalize original name
-            normalized_name = self._file_mapping.get(name, self._normalize_name(name))
-            obj = self.bucket.Object(normalized_name)
-
-            # Delete the object
-            obj.delete()
-
-            # Remove from mapping if exists
-            self._file_mapping.pop(name, None)
-
-            LOGGER.debug(
-                "File deleted from S3",
-                key=normalized_name,
-                tenant=self.tenant_prefix,
-            )
-        except ClientError as e:
-            error_code = e.response.get("Error", {}).get("Code")
-            if error_code not in ["404", "NoSuchKey"]:
-                LOGGER.error(
-                    "Failed to delete file from S3",
-                    name=name,
-                    error=str(e),
-                    tenant=self.tenant_prefix,
-                )
-                if error_code == "NoSuchBucket":
-                    raise S3BucketError(f"S3 bucket '{self._bucket_name}' does not exist") from e
-                elif error_code in ("AccessDenied", "AllAccessDisabled"):
-                    raise S3AccessError(f"No permission to access S3 bucket '{self._bucket_name}'") from e
-                else:
-                    raise S3StorageError(f"Failed to delete file from S3: {error_code}") from e
-            LOGGER.debug(
-                "File not found during delete",
-                name=name,
-                tenant=self.tenant_prefix,
-            )
-
-    def url(self, name: str, **kwargs) -> str:
-        """Generate URL for accessing the file.
-
-        Generates a signed URL for the file since buckets are private.
-        AWS signing parameters are required and preserved for authenticated access.
-
-        Args:
-            name (str): Name of the file
-            **kwargs: Additional arguments passed to the parent implementation
-
-        Returns:
-            str: Signed URL for accessing the file
-
-        Raises:
-            ClientError: If URL generation fails
-        """
-        try:
-            normalized_name = self._normalize_name(name)
-            LOGGER.debug(
-                "Generating URL for S3 object",
-                original_name=name,
-                normalized_name=normalized_name,
-                custom_domain=self._custom_domain,
-                endpoint_url=self._endpoint_url,
-                kwargs=kwargs,
-                tenant=self.tenant_prefix,
-            )
-
-            _ = self.client
-
-            # Generate presigned URL with explicit signing parameters
-            url = self._s3_client.generate_presigned_url(
-                "get_object",
-                Params={
-                    "Bucket": self._bucket_name,
-                    "Key": normalized_name,
-                    "ResponseContentDisposition": "inline",
-                },
-                ExpiresIn=3600,
-                HttpMethod="GET",
-            )
-
-            # If we have a custom domain, we need to preserve AWS signing parameters
-            if self._custom_domain:
-                try:
-                    # Parse the original URL to get AWS signing parameters
-                    parsed = urlparse(url)
-                    query_params = parse_qs(parsed.query)
-
-                    # Create new URL with custom domain but preserve AWS signing params
-                    custom_url = urlunparse(
-                        (
-                            "https" if CONFIG.get_bool("storage.media.s3.secure_urls", True) else "http",
-                            self._custom_domain,
-                            "/" + normalized_name,
-                            "",
-                            urlencode(query_params, doseq=True),  # Keep all AWS signing params
-                            "",
-                        )
-                    )
-
-                    LOGGER.debug(
-                        "Generated signed URL for custom domain",
-                        key=name,
-                        normalized_key=normalized_name,
-                        url=custom_url,
-                        custom_domain=self._custom_domain,
-                        has_aws_algorithm=bool(query_params.get("X-Amz-Algorithm")),
-                        has_aws_credential=bool(query_params.get("X-Amz-Credential")),
-                        has_aws_signature=bool(query_params.get("X-Amz-Signature")),
-                        tenant=self.tenant_prefix,
-                    )
-                    return custom_url
-                except ClientError as e:
-                    LOGGER.error(
-                        "Failed to generate signed URL",
-                        error_code=e.response["Error"]["Code"],
-                        message=e.response["Error"]["Message"],
-                        key=name,
-                        normalized_key=normalized_name,
-                        tenant=self.tenant_prefix,
-                    )
-                    raise
-            else:
-                LOGGER.debug(
-                    "Using standard S3 URL",
-                    name=normalized_name,
-                    url=url,
-                    has_aws_algorithm="X-Amz-Algorithm" in url,
-                    has_aws_credential="X-Amz-Credential" in url,
-                    has_aws_signature="X-Amz-Signature" in url,
-                    tenant=self.tenant_prefix,
-                )
-                return url
-
-        except ClientError as e:
-            LOGGER.error(
-                "S3 URL generation failed",
-                error_code=e.response["Error"]["Code"],
-                message=e.response["Error"]["Message"],
-                key=name,
-                tenant=self.tenant_prefix,
-            )
-            raise
-        except Exception as e:
-            LOGGER.error(
-                "Unexpected error generating URL",
-                name=name,
-                error=str(e),
-                tenant=self.tenant_prefix,
-            )
-            raise
-
-    def size(self, name: str) -> int:
-        """Get the size of a file.
-
-        Args:
-            name (str): Name of the file
-
-        Returns:
-            int: Size of the file in bytes
-        """
-        obj = self.bucket.Object(name)
-        return obj.content_length
-
-    def exists(self, name):
-        """Check if a file exists in S3.
-
-        Args:
-            name (str): Name of the file
-
-        Returns:
-            bool: True if file exists, False otherwise
-
-        Raises:
-            S3AccessError: If access is denied
-            S3BucketError: If bucket doesn't exist
-        """
-        tenant_prefixed_name = self._normalize_name(name)
-        LOGGER.debug(
-            "Checking if file exists", name=name, tenant_prefixed_name=tenant_prefixed_name
-        )
-
-        try:
-            # Use the S3 client to check if the object exists
-            self._s3_client.head_object(Bucket=self._bucket_name, Key=tenant_prefixed_name)
-            LOGGER.debug("File exists", name=name, tenant_prefixed_name=tenant_prefixed_name)
-            return True
-        except ClientError as e:
-            error_code = e.response.get("Error", {}).get("Code")
-            if error_code == "404":
-                LOGGER.debug(
-                    "File does not exist", name=name, tenant_prefixed_name=tenant_prefixed_name
-                )
-                return False
-            # Handle other client errors
-            LOGGER.error(
-                "Error checking if file exists",
-                name=name,
-                tenant_prefixed_name=tenant_prefixed_name,
-                error=str(e),
-                tenant=self.tenant_prefix,
-            )
-            if error_code == "NoSuchBucket":
-                raise S3BucketError(f"S3 bucket '{self._bucket_name}' does not exist") from e
-            elif error_code in ("AccessDenied", "AllAccessDisabled"):
-                raise S3AccessError(f"No permission to access S3 bucket '{self._bucket_name}'") from e
-            else:
-                raise S3StorageError(f"Error checking if file exists: {error_code}") from e
